Moved gitio test to separate file
